# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentação do melhor chatbot universitario",
        "title": "Golang CHATBOT",
        "contact": {},
        "version": "@0.0.1"
    },
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Autentica um usuário com email e senha",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Credenciais de login",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth_dto.LoginRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna um token JWT",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util_dto.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth_dto.AuthResponseDto"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Autentica um usuário com email e senha",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Cadastro de usuário",
                "parameters": [
                    {
                        "description": "Credenciais de login",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth_dto.RegisterRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna um token JWT",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util_dto.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth_dto.AuthResponseDto"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/chats/list-chats": {
            "get": {
                "description": "Lista chats aplicando filtros opcionais",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chats"
                ],
                "summary": "Listar chats",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "example": 10,
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "total",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de chats",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util_dto.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/chat_dto.ListChatDto"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Nenhum chat encontrado",
                        "schema": {
                            "$ref": "#/definitions/util_dto.AppResponse"
                        }
                    }
                }
            }
        },
        "/chats/new-chat": {
            "post": {
                "description": "Cria um novo chat com título e usuário associado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chats"
                ],
                "summary": "Criar novo chat",
                "parameters": [
                    {
                        "description": "Dados para criar chat",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat_dto.CreateChatDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat criado com sucesso",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util_dto.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/chat_dto.ChatDto"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Erro de validação ou requisição",
                        "schema": {
                            "$ref": "#/definitions/util_dto.AppResponse"
                        }
                    }
                }
            }
        },
        "/chats/{chatID}/": {
            "get": {
                "description": "Retorna os detalhes de um chat específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chats"
                ],
                "summary": "Buscar chat por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do chat",
                        "name": "chatID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat encontrado",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util_dto.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/chat_dto.ChatDto"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Nenhum chat encontrado",
                        "schema": {
                            "$ref": "#/definitions/util_dto.AppResponse"
                        }
                    }
                }
            }
        },
        "/chats/{chatID}/new-message": {
            "post": {
                "description": "Envia uma nova mensagem dentro de um chat existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chats"
                ],
                "summary": "Enviar mensagem para um chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do chat",
                        "name": "chatID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados da mensagem",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat_dto.CreateMensagemDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta da IA",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util_dto.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Erro de validação ou requisição",
                        "schema": {
                            "$ref": "#/definitions/util_dto.AppResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth_dto.AuthResponseDto": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/auth_dto.UserDto"
                }
            }
        },
        "auth_dto.LoginRequestDto": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth_dto.RegisterRequestDto": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth_dto.UserDto": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "chat_dto.ChatDto": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "chat_dto.CreateChatDto": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
        "chat_dto.CreateMensagemDto": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "chat_dto.ListChatDto": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/chat_dto.ChatDto"
                    }
                },
                "limit": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1,
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "util_dto.AppResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer Auth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}